+++ client/src/components/QA/components/QAnswer.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 function QAnswer({ answer }) {
   // console.log('body ', body);
-  console.log('answer:', answer);
+  // console.log('answer:', answer);
   return (
     <div>
       A: {answer.body}
