+++ client/src/components/RR/RRstyles/MoreButton.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const MoreButton = styled.button`
+  border: 2px solid black;
+  color: black;
+  padding: 15px 32px;
+  text-align: center;
+  display: inline-block;
+  font-size: 16px;
+`;
+
+export default MoreButton;
