+++ client/src/components/RR/RRstyles/RatingsBox.js
@@ -1,12 +1,13 @@
 import styled from 'styled-components';
 
 const RatingsBox = styled.div`
+  margin: 0;
+  padding: 0;
+  border: 0;
   height:200px;
   width:100%;
   background-color:#f5f9ff;
-  // align: center;
-  align-items: center;
-  // position: center;
+  text-align: center;
 `;
 
 export default RatingsBox;
