+++ client/src/components/RR/components/Ratings.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import RatingsBox from './RRstyles/RatingsBox.js';
+import RatingsBox from '../RRstyles/RatingsBox.js';
 
 function Ratings() {
   return (
     <RatingsBox>
-      <h5> avg num & some stuff </h5>
+      <h5> avg num & ratings </h5>
     </RatingsBox>
   );
 }
